options-infra/foundry-search/scripts/agents_v2/pyproject.toml
[project]
name = "agents-v2"
version = "0.1.0"
description = "Create Agents v2"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.13.3",
    "azure-ai-projects==2.0.0b2",
    "azure-identity>=1.25.0",
]
